+++ IPlug/IGraphics/IGraphicsMac.mm
@@ -161,7 +161,7 @@ bool IGraphicsMac::WindowIsOpen()
   return mView;
 }
 
-void IGraphicsMac::Resize(int w, int h, double scale)
+void IGraphicsMac::Resize(int w, int h, float scale)
 {
   if (w == Width() && h == Height() && scale == Scale()) return;
